pyproject.toml
[project]
name = "phd-welman"
version = "0.1.0"
description = "Repo for PhD work"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "jupyter>=1.1.1",
    "prompt-toolkit>=3.0.52",
    "pydantic>=2.12.5",
    "pydantic-ai>=1.43.0",
    "pydantic-graph>=1.43.0",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "rich>=14.2.0",
]

[build-system]
    requires = ["hatchling", "hatch-vcs"]
    build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
    packages = ["phd_welman"]
